\octave, 6 in the line 122 does not work
The variable melody uses midonote. Thus, \octave does not work.
If you want to show the usages of \octave, then you must change variable a and variable b.
If you want to use midinote then the keyword \octave must be got rid of, and 12 must be added as shown in my suggestion.